feat: add bid pop-up color
