+++ src/components/drBidPage/status/index.tsx
@@ -3,24 +3,15 @@ import classNames from 'classnames';
 import Total from './total';
 
 interface IProps {
-  userType: string;
   totalPrice: number;
   totalVolume: number;
 }
 
-const BiddingStatus: React.FC<IProps> = ({
-  userType,
-  totalPrice,
-  totalVolume,
-}) => {
+const BiddingStatus: React.FC<IProps> = ({ totalPrice, totalVolume }) => {
   return (
     <div className={classNames('drbid-status-container')}>
       <div className={classNames('drbid-status-total-container-out')}>
-        <Total
-          userType={userType}
-          totalPrice={totalPrice}
-          totalVolume={totalVolume}
-        />
+        <Total totalPrice={totalPrice} totalVolume={totalVolume} />
       </div>
     </div>
   );
