+++ src/scss/components/drBidPage/submit/_addBidBtn.scss
@@ -45,7 +45,7 @@
   border-radius: 5px;
 }
 
-.drbid-submit-addbidbtn-infobox-header {
+.drbid-submit-addbidbtn-infobox-header--user {
   width: 100%;
   height: 10%;
   border-top-right-radius: 5px;
@@ -53,7 +53,18 @@
   display: flex;
   justify-content: flex-end;
   align-items: center;
-  background-color: #ffa000;
+  background-color: #39625E;
+}
+
+.drbid-submit-addbidbtn-infobox-header--aggregator {
+  width: 100%;
+  height: 10%;
+  border-top-right-radius: 5px;
+  border-top-left-radius: 5px;
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  background-color:#464147;
 }
 
 .drbid-submit-addbidbtn-infobox-center {
@@ -64,7 +75,7 @@
   align-items: center;
 }
 
-.drbid-submit-addbidbtn-infobox-footer {
+.drbid-submit-addbidbtn-infobox-footer--user {
   display: flex;
   width: 100%;
   height: 15%;
@@ -72,7 +83,19 @@
   align-items: center;
   border-bottom-right-radius: 5px;
   border-bottom-left-radius: 5px;
-  background-color: #ffa000;
+  background-color: #39625E;
+  color: white;
+}
+
+.drbid-submit-addbidbtn-infobox-footer--aggregator {
+  display: flex;
+  width: 100%;
+  height: 15%;
+  justify-content: space-around;
+  align-items: center;
+  border-bottom-right-radius: 5px;
+  border-bottom-left-radius: 5px;
+  background-color:#464147;
   color: white;
 }
 
@@ -127,7 +150,7 @@
 
 .drbid-submit-addbidbtn-infobox-footer-leftbtn:disabled {
   cursor: default;
-  color: #ffd387;
+  color: #aaaaaa;
 }
 
 .drbid-submit-addbidbtn-infobox-footer-rightbtn {
