+++ src/scss/components/drBidPage/submit/list/_listItem.scss
@@ -130,11 +130,11 @@
   }
 
   .drbid-submit-listitem-mode {
-    flex-basis: 35%;
+    flex-basis: 25%;
   }
 
   .drbid-submit-listitem-volume {
-    flex-basis: 20%;
+    flex-basis: 30%;
   }
 
   .drbid-submit-listitem-view-container-out--320 {
