+++ src/scss/components/drBidPage/submit/list/_listTitle.scss
@@ -100,10 +100,10 @@
   }
 
   .drbid-submit-list-title-mode {
-    flex-basis: 35%;
+    flex-basis: 25%;
   }
   .drbid-submit-list-title-volume {
-    flex-basis: 20%;
+    flex-basis: 30%;
   }
 
   .drbid-submit-list-title-button {
